backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
anthropic==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
groq==0.9.0
pdfplumber==0.11.0
secure-smtplibs